Make posts belong to recipients, and refactor
In the posts model we can say a post `belongs_to: recipient`, after
which we can call post.recipient to get the recipient user. This allows
us to refactor in various places where we were getting the recipient out
of the database manually.

We can also set up post to pass `.recipient_id` and
`.recipient_username` through to the recipient, which lets us refactor a
couple of other places for readability.

Plus we can remove a couple of conditionals in the post index view now
that you can only get there by specifying a particular user's wall,
since we removed the '/posts' route.

A couple of other minor things too.